requirements.txt
amqp==1.4.9
elasticsearch2==2.5.0
facebook-business==3.2.3
mysqlclient==1.3.7
oslo.config==6.2.1
oslo.context==2.20.0
oslo.db==4.20.0
oslo.i18n==3.20.0
oslo.log==3.38.1
oslo.serialization==2.20.0
oslo.utils==3.36.2
paho-mqtt==1.3.1
pika==0.10.0
pymongo==3.6.1
PyMySQL==0.7.11
pyparsing==2.2.0
redis==2.10.5
requests==2.18.4
SQLAlchemy==1.2.0
tornado==4.5.3
